Add tests for item-id route and consolidate tests in one file
