--- sabelotodo/tests/test_backend.py
@@ -1,6 +1,0 @@
-import json
-
-
-def test_empty_db(test_client, _db):
-    rv = test_client.get('/item')
-    assert json.loads(rv.data) == []
