--- sabelotodo/tests/test_hello.py
@@ -1,14 +1,0 @@
-from sabelotodo.models import Item
-
-
-def test_create_item(_db):
-    """ Verify that database works"""
-    name = 'TestItem'
-    done = True
-    i = Item(name=name, order=0, done=done)
-
-    _db.session.add(i)
-    _db.session.commit()
-
-    item = Item.query.filter_by(name=name).first()
-    assert item.done == done
